Added hooks for hwdep read/write access.
This interface is intended to allow a user-space application to be able
to select from a number of different signal sources whilst the driver is
'live' and potentially in use.

Also added a hook for the proc 'info' entry. This will eventually give a
lot more details about the driver.